src/requirements.txt
Django>=3.2.15
gunicorn>=20.1.0
django-multiselectfield>=0.1.12
django-crispy-forms>=1.14.0
crispy-bootstrap5>=0.7
envconfiguration>=0.1.0
mysqlclient>=2.1.1
django-select2>=7.10.1
django_redis>=5.2.0
envconfiguration>=0.1.0
django-tables2>=2.4.1
django-filter>=22.1
django-bootstrap-v5>=1.0.11
django_bootstrap_icons>=0.8.1
django-fontawesome-5>=1.0.18
tablib>=3.2.1
tablib[xls]>=3.2.1
tablib[xlsx]>=3.2.1
